+++ sdk/eventgrid/azure-mgmt-eventgrid/generated_samples/domains_delete.py
@@ -35,6 +35,6 @@ def main():
     ).result()
 
 
-# x-ms-original-file: specification/eventgrid/resource-manager/Microsoft.EventGrid/preview/2023-06-01-preview/examples/Domains_Delete.json
+# x-ms-original-file: specification/eventgrid/resource-manager/Microsoft.EventGrid/preview/2023-12-15-preview/examples/Domains_Delete.json
 if __name__ == "__main__":
     main()