+++ sdk/eventgrid/azure-mgmt-eventgrid/generated_samples/namespace_topic_event_subscriptions_list_by_namespace_topic.py
@@ -38,6 +38,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: specification/eventgrid/resource-manager/Microsoft.EventGrid/preview/2023-06-01-preview/examples/NamespaceTopicEventSubscriptions_ListByNamespaceTopic.json
+# x-ms-original-file: specification/eventgrid/resource-manager/Microsoft.EventGrid/preview/2023-12-15-preview/examples/NamespaceTopicEventSubscriptions_ListByNamespaceTopic.json
 if __name__ == "__main__":
     main()